adding optimizer to opt_result
