+++ qadence/ml_tools/config.py
@@ -13,7 +13,7 @@ from torch import Tensor
 
 from qadence.blocks.analog import AnalogBlock
 from qadence.blocks.primitive import ParametricBlock
-from qadence.ml_tools.utils import OptimizeResult
+from qadence.ml_tools.data import OptimizeResult
 from qadence.operations import RX, AnalogRX
 from qadence.parameters import Parameter
 from qadence.types import (
@@ -44,6 +44,8 @@ class Callback:
             If callback_condition is None, we set
             callback_condition to returns True when iteration % every == 0
         call_before_opt (bool): If true, callback is done before training.
+            Note, this means this applies only on the validation data as
+            we only perform validation before training.
     """
 
     def __init__(
